requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.14.3
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.19.2
openpyxl==3.0.5
pandas==1.1.3
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
selenium==3.141.0
six==1.15.0
urllib3==1.25.10
xlrd==1.2.0
